Move PersistentVolumeClaim logic to its own module
