+++ pkg/controller/controller.go
@@ -203,28 +203,6 @@ func (hc *HabitatController) cacheConfigMaps() {
 	hc.cmInformerSynced = hc.cmInformer.HasSynced
 }
 
-func (hc *HabitatController) cachePersistentVolumeClaims() {
-	source := newListWatchFromClientWithLabels(
-		hc.config.KubernetesClientset.CoreV1().RESTClient(),
-		"persistentvolumeclaims",
-		apiv1.NamespaceAll,
-		labelListOptions())
-
-	hc.cmInformer = cache.NewSharedIndexInformer(
-		source,
-		&apiv1.PersistentVolumeClaim{},
-		resyncPeriod,
-		cache.Indexers{},
-	)
-
-	hc.cmInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-		UpdateFunc: hc.handlePVCUpdate,
-		DeleteFunc: hc.handlePVCDelete,
-	})
-
-	hc.cmInformerSynced = hc.pvcInformer.HasSynced
-}
-
 func (hc *HabitatController) watchPods(ctx context.Context) {
 	source := newListWatchFromClientWithLabels(
 		hc.config.KubernetesClientset.CoreV1().RESTClient(),
@@ -315,46 +293,6 @@ func (hc *HabitatController) handleCMUpdate(oldObj, newObj interface{}) {
 
 func (hc *HabitatController) handleCMDelete(obj interface{}) {
 	hc.handleCM(obj)
-}
-
-func (hc *HabitatController) checkPVC(pvc *apiv1.PersistentVolumeClaim) {
-	// Find out if there's a Habitat object around
-	// If so, we have a problem.
-	// TODO we might use an ownerref to find the STS here
-	key := fmt.Sprintf("%s/%s", pvc.Namespace, pvc.Labels[habv1beta1.HabitatNameLabel])
-
-	_, exists, err := hc.habInformer.GetStore().GetByKey(key)
-	if err != nil {
-		level.Error(hc.logger).Log("msg", "Failed to get key in Store", "obj", key)
-		return
-	} else if !exists {
-		level.Debug(hc.logger).Log("msg", "No matching Habitat found for PVC", "name", pvc.Name)
-		return
-	} else {
-		level.Error(hc.logger).Log("msg", "A PVC has lost its PersistentVolume", "name", pvc.Name)
-	}
-}
-
-func (hc *HabitatController) handlePVCUpdate(oldObj, newObj interface{}) {
-	pvc, ok := newObj.(*apiv1.PersistentVolumeClaim)
-	if !ok {
-		level.Error(hc.logger).Log("msg", "Failed to type assert PersistentVolumeClaim", "obj", newObj)
-		return
-	}
-
-	if pvc.Status.Phase == apiv1.ClaimLost {
-		hc.checkPVC(pvc)
-	}
-}
-
-func (hc *HabitatController) handlePVCDelete(obj interface{}) {
-	pvc, ok := obj.(*apiv1.PersistentVolumeClaim)
-	if !ok {
-		level.Error(hc.logger).Log("msg", "Failed to type assert PersistentVolumeClaim", "obj", obj)
-		return
-	}
-
-	hc.checkPVC(pvc)
 }
 
 func (hc *HabitatController) handlePodAdd(obj interface{}) {
